+++ src/pages/BPMN_generator.tsx
@@ -17,6 +17,7 @@
 import React from 'react';
 import { Heading, Text } from 'rebass/styled-components';
 
+import { Jotform } from '../theme/components/Jotform';
 import Section from '../theme/components/Section';
 import Triangle from '../theme/components/Triangle';
 import Footer from '../theme/components/Footer';
@@ -27,146 +28,8 @@ import { PAGE } from '../theme/utils/constants';
 // TODO duplicated from Landing.tsx
 const centerHorizontally = { marginRight: 'auto', marginLeft: 'auto' };
 
-/* eslint-disable @typescript-eslint/no-explicit-any */
-const handleIFrameMessage = function (e: any): void {
-  if (typeof e.data === 'object') {
-    return;
-  }
-  const args = e.data.split(':');
-  const iframe: HTMLIFrameElement = document.getElementById(
-    args.length > 2
-      ? 'JotFormIFrame-' + args[args.length - 1]
-      : 'JotFormIFrame',
-  ) as HTMLIFrameElement;
-
-  if (!iframe) {
-    return;
-  }
-
-  switch (args[0]) {
-    case 'scrollIntoView':
-      iframe.scrollIntoView();
-      break;
-    case 'setHeight':
-      iframe.style.height = args[1] + 'px';
-      break;
-    case 'collapseErrorPage':
-      if (iframe.clientHeight > window.innerHeight) {
-        iframe.style.height = window.innerHeight + 'px';
-      }
-      break;
-    case 'reloadPage':
-      window.location.reload();
-      break;
-    case 'loadScript':
-      if (!isPermitted(e.origin, ['jotform.com', 'jotform.pro'])) {
-        break;
-      }
-      let src = args[1];
-      if (args.length > 3) {
-        src = args[1] + ':' + args[2];
-      }
-      const script = document.createElement('script');
-      script.src = src;
-      script.type = 'text/javascript';
-      document.body.appendChild(script);
-      break;
-    case 'exitFullscreen':
-      if (window.document.exitFullscreen) window.document.exitFullscreen();
-      /*      else if (window.document.mozCancelFullScreen)
-        window.document.mozCancelFullScreen();
-      else if (window.document.mozCancelFullscreen)
-        window.document.mozCancelFullScreen();
-      else if (window.document.webkitExitFullscreen)
-        window.document.webkitExitFullscreen();
-      else if (window.document.msExitFullscreen)
-        window.document.msExitFullscreen();*/
-      break;
-  }
-  const isJotForm = e.origin.indexOf('jotform') > -1 ? true : false;
-
-  if (isJotForm && iframe.contentWindow?.postMessage) {
-    const urls = {
-      docurl: encodeURIComponent(document.URL),
-      referrer: encodeURIComponent(document.referrer),
-    };
-    iframe.contentWindow.postMessage(
-      JSON.stringify({ type: 'urls', value: urls }),
-      '*',
-    );
-  }
-};
-
-const isPermitted = function (
-  originUrl: string,
-  whitelisted_domains: any,
-): boolean {
-  const url = document.createElement('a');
-  url.href = originUrl;
-  const hostname = url.hostname;
-  let result = false;
-  if (typeof hostname !== 'undefined') {
-    whitelisted_domains.forEach(function (element: any): void {
-      if (
-        hostname.slice(-1 * element.length - 1) === '.'.concat(element) ||
-        hostname === element
-      ) {
-        result = true;
-      }
-    });
-    return result;
-  }
-  return result;
-};
-
 const BPMNGeneratorPage = (): JSX.Element => {
-  const ifr: React.DetailedHTMLProps<
-    React.IframeHTMLAttributes<HTMLIFrameElement>,
-    HTMLIFrameElement
-  > = (
-    <iframe
-      id="JotFormIFrame-221232749309354"
-      title="BPMN generation from event logs"
-      onLoad={() => window.parent.scrollTo(0, 0)}
-      allowTransparency={true}
-      allowFullScreen={true}
-      allow="geolocation; microphone; camera"
-      src="https://form.jotform.com/221232749309354"
-      frameBorder="0"
-      style={{
-        border: 'none',
-        minWidth: '25%',
-        width: '30vw',
-        height: '75vh',
-        // margin: '1rem',
-      }}
-      scrolling="no"
-    ></iframe>
-  );
-
-  let src = ifr.src;
-  let iframeParams: string[] = [];
-  if (window.location.href && window.location.href.indexOf('?') > -1) {
-    iframeParams = iframeParams.concat(
-      window.location.href
-        .substr(window.location.href.indexOf('?') + 1)
-        .split('&'),
-    );
-  }
-  if (src && src.indexOf('?') > -1) {
-    iframeParams = iframeParams.concat(
-      src.substr(src.indexOf('?') + 1).split('&'),
-    );
-    src = src.substr(0, src.indexOf('?'));
-  }
-  iframeParams.push('isIframeEmbed=1');
-  ifr.src = src + '?' + iframeParams.join('&');
-
-  if (window.addEventListener) {
-    window.addEventListener('message', handleIFrameMessage, false);
-    /*  } else if (window.attachEvent) {
-    window.attachEvent('onmessage', handleIFrameMessage);*/
-  }
+  const ifr = Jotform();
 
   return (
     <Layout title={PAGE.bpmn_generator}>
